+++ src/kairos/authentication/application/login.test.ts
@@ -1,10 +1,10 @@
-import { describe, expect, it, vi } from 'vitest';
-import { type GenerateAccessToken, createLogin, type VerifyPassword } from '@/kairos/authentication/application/login';
+import { createLogin, type GenerateAccessToken, type VerifyPassword } from '@/kairos/authentication/application/login';
 import { type AccessToken } from '@/kairos/authentication/domain/access-token';
 import { invalidCredentialsError } from '@/kairos/authentication/domain/errors';
 import type { User } from '@/kairos/authentication/domain/user';
 import { type FindOneByEmail } from '@/kairos/authentication/domain/user-credentials-repository';
-import { err, ok } from '@/kairos/shared/kernel/result';
+import { err, ok } from '@/kairos/shared/result/result';
+import { describe, expect, it, vi } from 'vitest';
 
 describe('Authenticate use case', () => {
   const user: User = {
