+++ src/kairos/authentication/domain/user-credentials-repository.ts
@@ -1,5 +1,5 @@
 import { type InvalidCredentialsError } from '@/kairos/authentication/domain/errors';
 import type { User } from '@/kairos/authentication/domain/user';
-import { type Result } from '@/kairos/shared/kernel/result';
+import { type Result } from '@/kairos/shared/result/result';
 
 export type FindOneByEmail = (email: string) => Promise<Result<User, InvalidCredentialsError>>;
