+++ src/kairos/setup/application/ensure-super-admin.ts
@@ -1,10 +1,10 @@
+import { type SetupTask } from '@/kairos/setup/application/run-setup';
 import {
   type CreateSuperAdmin,
   type ExistsSuperAdmin,
   type SuperAdminCredentials,
 } from '@/kairos/setup/domain/super-admin-repository';
-import { ok, type Result } from '@/kairos/shared/kernel/result';
-import { type SetupTask } from '@/kairos/setup/application/run-setup';
+import { ok, type Result } from '@/kairos/shared/result/result';
 
 type EnsureSuperAdminDependencies = Readonly<{
   existsSuperAdmin: ExistsSuperAdmin;
