+++ src/kairos/shared/result/result.ts
